Switched all timers to 64-bit unsigned integers.

+++ SA2 Battle Network/MemoryHandler.h
@@ -104,7 +104,7 @@ private:
 
 	// Analog throttle timer.
 	// Prevents it from spamming packets.
-	uint analogTimer;
+	uint64 analogTimer;
 
 	// Toggles and things
 	bool firstMenuEntry;
